backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
tldextract>=5.1.0
slowapi>=0.1.9
python-dotenv>=1.0.0
